requirements.txt
# Virtualenv with system libraries since SimpleCV doesn't
# work well under virtualenv
Flask==0.10.1
Flask-Bootstrap==3.1.1.2
Flask-SocketIO==0.3.8
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.3
PIL==1.1.7
RPi.GPIO==0.5.6
SimpleCV==1.3
Werkzeug==0.9.6
argparse==1.2.1
astroid==1.1.1
configobj==4.7.2
decorator==3.3.3
distribute==0.6.24
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.2
ipython==0.13.1
itsdangerous==0.24
jsonschema==2.3.0
lockfile==0.9.1
logilab-common==0.62.0
nose==1.3.3
numpy==1.6.2
pep8==1.5.7
pexpect==2.4
pifacecommon==4.0.0
pifacedigitalio==3.0.4
pycurl==7.19.0
pygame==1.9.1release
pylint==1.2.1
pyparsing==2.0.2
python-daemon==1.6
pyzmq==14.3.1
scipy==0.10.1
serpent==1.7
simplegeneric==0.8.1
svgwrite==1.1.6
tornado==2.3
ufw==0.31.1-2
virtualenv==1.7.1.2
wsgiref==0.1.2
